+++ plugin/android-junit5/src/main/templates/Libraries.kt
@@ -1,6 +1,8 @@
 package de.mannodermaus
 
 internal object Libraries {
-    const val instrumentationCore = "@INSTRUMENTATION_GROUP@:@INSTRUMENTATION_CORE@:@INSTRUMENTATION_VERSION@"
-    const val instrumentationRunner = "@INSTRUMENTATION_GROUP@:@INSTRUMENTATION_RUNNER@:@INSTRUMENTATION_VERSION@"
+    const val instrumentationVersion = "@INSTRUMENTATION_VERSION@"
+    const val instrumentationCore = "@INSTRUMENTATION_GROUP@:@INSTRUMENTATION_CORE@"
+    const val instrumentationExtensions = "@INSTRUMENTATION_GROUP@:@INSTRUMENTATION_EXTENSIONS@"
+    const val instrumentationRunner = "@INSTRUMENTATION_GROUP@:@INSTRUMENTATION_RUNNER@"
 }